KIALI-2010 Remove 'containers' folder
Moving container components to be on the same source file as
the containerized components.
