+++ src/components/EmptyGraphLayout.tsx
