+++ src/components/MessageCenter/MessageCenterTrigger.tsx
@@ -1,6 +1,12 @@
 import * as React from 'react';
+import { ThunkDispatch } from 'redux-thunk';
+import { connect } from 'react-redux';
 import { Badge, Button, ButtonVariant } from '@patternfly/react-core';
 import { BellIcon, WarningTriangleIcon } from '@patternfly/react-icons';
+import { KialiAppState } from '../../store/Store';
+import { MessageType, NotificationGroup, NotificationMessage } from '../../types/MessageCenter';
+import { KialiAppAction } from '../../actions/KialiAppAction';
+import MessageCenterThunkActions from '../../actions/MessageCenterThunkActions';
 
 type PropsType = {
   newMessagesCount: number;
@@ -10,7 +16,7 @@ type PropsType = {
   toggleSystemErrorsCenter: () => void;
 };
 
-export default class MessageCenterTrigger extends React.PureComponent<PropsType, {}> {
+export class MessageCenterTrigger extends React.PureComponent<PropsType, {}> {
   render() {
     return (
       <>
@@ -57,3 +63,52 @@ export default class MessageCenterTrigger extends React.PureComponent<PropsType,
     );
   };
 }
+
+const mapStateToPropsMessageCenterTrigger = (state: KialiAppState) => {
+  type MessageCenterTriggerPropsToMap = {
+    newMessagesCount: number;
+    badgeDanger: boolean;
+    systemErrorsCount: number;
+  };
+
+  const dangerousMessageTypes = [MessageType.ERROR, MessageType.WARNING];
+  let systemErrorsCount = 0;
+
+  const systemErrorsGroup = state.messageCenter.groups.find(item => item.id === 'systemErrors');
+  if (systemErrorsGroup) {
+    systemErrorsCount = systemErrorsGroup.messages.length;
+  }
+
+  return state.messageCenter.groups
+    .reduce((unreadMessages: NotificationMessage[], group: NotificationGroup) => {
+      return unreadMessages.concat(
+        group.messages.reduce((unreadMessagesInGroup: NotificationMessage[], message: NotificationMessage) => {
+          if (!message.seen) {
+            unreadMessagesInGroup.push(message);
+          }
+          return unreadMessagesInGroup;
+        }, [])
+      );
+    }, [])
+    .reduce(
+      (propsToMap: MessageCenterTriggerPropsToMap, message: NotificationMessage) => {
+        propsToMap.newMessagesCount++;
+        propsToMap.badgeDanger = propsToMap.badgeDanger || dangerousMessageTypes.includes(message.type);
+        return propsToMap;
+      },
+      { newMessagesCount: 0, systemErrorsCount: systemErrorsCount, badgeDanger: false }
+    );
+};
+
+const mapDispatchToPropsMessageCenterTrigger = (dispatch: ThunkDispatch<KialiAppState, void, KialiAppAction>) => {
+  return {
+    toggleMessageCenter: () => dispatch(MessageCenterThunkActions.toggleMessageCenter()),
+    toggleSystemErrorsCenter: () => dispatch(MessageCenterThunkActions.toggleSystemErrorsCenter())
+  };
+};
+
+const MessageCenterTriggerContainer = connect(
+  mapStateToPropsMessageCenterTrigger,
+  mapDispatchToPropsMessageCenterTrigger
+)(MessageCenterTrigger);
+export default MessageCenterTriggerContainer;
