+++ src/components/Nav/Masthead/Masthead.tsx
@@ -1,18 +1,18 @@
 import * as React from 'react';
 
-import PfSpinnerContainer from '../../../containers/PfSpinnerContainer';
+import PfSpinner from '../../PfSpinner';
 import { Toolbar as ToolbarNext, ToolbarGroup, ToolbarItem } from '@patternfly/react-core';
 import UserDropdown from './UserDropdown';
 import { default as MeshMTLSStatus } from '../../../components/MTls/MeshMTLSStatus';
 import HelpDropdown from './HelpDropdown';
-import { MessageCenterTriggerContainer } from '../../../containers/MessageCenterContainer';
+import MessageCenterTriggerContainer from '../../../components/MessageCenter/MessageCenterTrigger';
 
 class Masthead extends React.Component {
   render() {
     return (
       <ToolbarNext>
         <ToolbarGroup>
-          <PfSpinnerContainer />
+          <PfSpinner />
         </ToolbarGroup>
         <ToolbarGroup>
           <ToolbarItem>
